moved `goToNearestFuncDef` to `GoToFuncId`
