+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/search/gotodef/GoToFuncId.scala
@@ -20,12 +20,13 @@ import org.alephium.protocol.vm.StatefulContext
 import org.alephium.ralph.Ast
 import org.alephium.ralph.lsp.access.compiler.ast.AstExtra
 import org.alephium.ralph.lsp.access.compiler.ast.node.Node
+import org.alephium.ralph.lsp.access.compiler.message.SourceIndexExtra.SourceIndexExtension
 import org.alephium.ralph.lsp.pc.log.{ClientLogger, StrictImplicitLogging}
 import org.alephium.ralph.lsp.pc.sourcecode.SourceLocation
 import org.alephium.ralph.lsp.pc.workspace.{WorkspaceState, WorkspaceSearcher}
 import org.alephium.ralph.lsp.pc.workspace.build.dependency.DependencyID
 
-private object GoToFuncId extends StrictImplicitLogging {
+private[search] object GoToFuncId extends StrictImplicitLogging {
 
   /**
    * Navigate to the definition of a function for the given [[Ast.FuncId]].
@@ -90,6 +91,34 @@ private object GoToFuncId extends StrictImplicitLogging {
 
       case None =>
         Iterator.empty
+    }
+
+  /**
+   * Navigate to the nearest function definition for which the given child node is in scope.
+   *
+   * @param childNode The node to traverse up from.
+   * @return An Option containing the nearest function definition, if found.
+   */
+  def goToNearestFuncDef(childNode: Node[Ast.Positioned, Ast.Positioned]): Option[Node[Ast.FuncDef[_], Ast.Positioned]] =
+    childNode.data match {
+      case function: Ast.FuncDef[_] =>
+        // Nested function definitions are not allowed in Ralph.
+        // If the input node is a function, return the node itself.
+        Some(childNode.upcast(function))
+
+      case ast: Ast.Positioned =>
+        // For everything else, find the nearest function.
+        ast
+          .sourceIndex
+          .flatMap {
+            childNodeIndex =>
+              childNode
+                .walkParents
+                .collectFirst {
+                  case node @ Node(function: Ast.FuncDef[_], _) if function.sourceIndex.exists(_ contains childNodeIndex.index) =>
+                    node.upcast(function)
+                }
+          }
     }
 
   /**
