+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/search/gotodef/GoToIdent.scala
@@ -19,7 +19,6 @@ package org.alephium.ralph.lsp.pc.search.gotodef
 import org.alephium.ralph.Ast
 import org.alephium.ralph.lsp.access.compiler.ast.node.Node
 import org.alephium.ralph.lsp.access.compiler.ast.{AstExtra, Tree}
-import org.alephium.ralph.lsp.access.compiler.message.SourceIndexExtra._
 import org.alephium.ralph.lsp.pc.sourcecode.SourceLocation
 import org.alephium.ralph.lsp.pc.workspace.{WorkspaceState, WorkspaceSearcher}
 
@@ -205,7 +204,7 @@ private[search] object GoToIdent {
       argumentNode: Node[Ast.Argument, Ast.Positioned],
       sourceCode: SourceLocation.Code,
       workspace: WorkspaceState.IsSourceAware): Iterator[SourceLocation.Node[Ast.Ident]] =
-    goToNearestFuncDef(argumentNode) match {
+    GoToFuncId.goToNearestFuncDef(argumentNode) match {
       case Some(functionNode) =>
         // It's a function argument, search within this function's body.
         goToVariableUsages(
@@ -428,38 +427,10 @@ private[search] object GoToIdent {
     sourceTree.ast match {
       case Left(_: Ast.Contract | _: Ast.ContractInterface | _: Ast.TxScript) =>
         // Find the nearest function definition or use the template body as the scope.
-        goToNearestFuncDef(childNode).orElse(Some(sourceTree.rootNode))
+        GoToFuncId.goToNearestFuncDef(childNode).orElse(Some(sourceTree.rootNode))
 
       case Right(_: Ast.Struct) =>
         None
-    }
-
-  /**
-   * Navigate to the nearest function definition for which the given child node is in scope.
-   *
-   * @param childNode The node to traverse up from.
-   * @return An Option containing the nearest function definition, if found.
-   */
-  def goToNearestFuncDef(childNode: Node[Ast.Positioned, Ast.Positioned]): Option[Node[Ast.FuncDef[_], Ast.Positioned]] =
-    childNode.data match {
-      case function: Ast.FuncDef[_] =>
-        // Nested function definitions are not allowed in Ralph.
-        // If the input node is a function, return the node itself.
-        Some(childNode.upcast(function))
-
-      case ast: Ast.Positioned =>
-        // For everything else, find the nearest function.
-        ast
-          .sourceIndex
-          .flatMap {
-            childNodeIndex =>
-              childNode
-                .walkParents
-                .collectFirst {
-                  case node @ Node(function: Ast.FuncDef[_], _) if function.sourceIndex.exists(_ contains childNodeIndex.index) =>
-                    node.upcast(function)
-                }
-          }
     }
 
   /**
@@ -469,7 +440,7 @@ private[search] object GoToIdent {
    * @return An array sequence of [[Ast.Argument]]s matching the search result.
    */
   private def goToNearestFunctionArguments(identNode: Node[Ast.Ident, Ast.Positioned]): Iterator[Ast.Argument] =
-    goToNearestFuncDef(identNode) match {
+    GoToFuncId.goToNearestFuncDef(identNode) match {
       case Some(functionNode) =>
         functionNode
           .data
